requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
yt-dlp
python-dotenv==1.0.1
